requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.6
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7
PyPDF2==3.0.1
Pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3
numpy==1.24.3
camelot-py==0.11.0
opencv-python==4.8.1.78
pdf2image==1.16.3
pytesseract==0.3.10
python-dotenv==1.0.0
ghostscript==0.7 